feat(multiple): add an option to set token renewal values for shared access key authentication
Added the option at the device client level.  Adjusted the transports to validate the options.
Adjusted the SAK authentication provider to have a new method to set the renewal values.  If the
authentication provider is NOT already "running" then just set the values.  Otherwise set the values
and cause a new token event available to be emitted.